DolphinQt/Config: Resolve sign-compare warnings

+++ Source/Core/DolphinQt/Config/GeckoCodeWidget.cpp
@@ -471,7 +471,7 @@ void GeckoCodeWidget::ToggleAllCodes()
     // file once per code.
     QSignalBlocker blocker(m_code_list);
 
-    for (int i = 0; i < m_gecko_codes.size(); ++i)
+    for (int i = 0; i < static_cast<int>(m_gecko_codes.size()); ++i)
     {
       m_gecko_codes[i].enabled = new_state;
       m_code_list->item(i)->setCheckState(new_check_state);
